Speed up parallelism (omp_cores + double for loop)
Two interrelated changes to speed up RCTD parallelism
- convert the main inner foreach::foreach loops into a 2D loop, where it first
splits up the array into smaller chunks, then for loops with*in* those chunks.
In preliminary tests, this gives speedups because the innerloop iterated so fast
and foreach had some overhead
- limit with omp_set_num_threads and blas_set_num_threads to a smaller number.
The default matmult in R uses blas/omp with lots of parallelism but this was
overloading the cores and for the small matrices RCTD uses, was inefficient,
especially on many-core machines.
Forcing these number of threads to be much smaller (1-5) allowed greater and
faster parallelism with less core contention